backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.32.1
pydantic-settings==2.6.1
pydantic==2.10.3
supabase==2.10.0
python-jose[cryptography]==3.3.0
redis[asyncio]==5.2.1
celery==5.4.0
slowapi==0.1.9
httpx
pytest==8.3.4
pytest-asyncio==0.24.0
